Coupling fix + code refactor
